requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
flake8==5.0.4
greenlet==1.1.3.post0
idna==3.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.42
tqdm==4.64.1
urllib3==1.26.12
